Some clean up.
