+++ include/Async/libuv.h
@@ -20,11 +20,6 @@ namespace clice::async {
 /// The default event loop.
 extern uv_loop_t* loop;
 
-#define uv_check_call(func, ...)                                                                   \
-    if(int error = func(__VA_ARGS__); error < 0) {                                                 \
-        log::fatal("An error occurred in " #func ": {0}", uv_strerror(error));                     \
-    }
-
 template <typename T, typename U>
 T& uv_cast(U* u) {
     assert(u && u->data && "uv_cast: invalid uv handle");
@@ -32,7 +27,5 @@ T& uv_cast(U* u) {
 }
 
 const std::error_category& category();
-
-
 
 }  // namespace clice::async
