+++ src/Driver/integration_tests.cc
@@ -1,5 +1,5 @@
 #include "Async/Async.h"
-#include "Server/Logger.h"
+#include "Support/Logger.h"
 #include "Support/Support.h"
 #include "llvm/Support/CommandLine.h"
 
@@ -46,13 +46,12 @@ int main(int argc, const char** argv) {
     llvm::cl::SetVersionPrinter([](llvm::raw_ostream& os) { os << "clice version: 0.0.1\n"; });
     llvm::cl::ParseCommandLineOptions(argc, argv, "clice language server");
 
-    async::net::spawn(
-        [](json::Value value) -> async::Task<> {
-            print("Receive: {0}", value);
-            co_return;
-        },
-        cl::execute.getValue(),
-        {"--pipe=true", "--config=/home/ykiko/C++/clice2/docs/clice.toml"});
+    async::net::spawn(cl::execute.getValue(),
+                      {"--pipe=true", "--config=/home/ykiko/C++/clice2/docs/clice.toml"},
+                      [](json::Value value) -> async::Task<> {
+                          print("Receive: {0}", value);
+                          co_return;
+                      });
 
     auto p = request("initialize", "input.json", "initialize");
     async::run(p);
