+++ src/Server/Database.cpp
@@ -1,4 +1,4 @@
-#include "Server/Logger.h"
+#include "Support/Logger.h"
 #include "Server/Database.h"
 #include "Support/Support.h"
 #include "Compiler/Compilation.h"
